requirements.txt
aiosmtpd==1.4.6
asgiref==3.8.1
atpublic==5.0
attrs==24.2.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
nodeenv==1.9.1
PyJWT==2.10.1
pyright==1.1.389
sqlparse==0.5.2
typing_extensions==4.12.2
